First draft workspace system
