+++ packages/framework/esm-styleguide/src/workspaces/public.ts
@@ -1,0 +1,3 @@
+export { launchWorkspace } from './workspaces';
+export { WorkspaceWindow } from './window/workspace-window.component';
+export { WorkspaceOverlay } from './overlay/workspace-overlay.component';
