+++ packages/framework/esm-translations/src/translations.ts
@@ -19,6 +19,7 @@ export const coreTranslations = {
   cancel: 'Cancel',
   change: 'Change',
   close: 'Close',
+  closeWorkspace: 'Close workspace',
   confirm: 'Confirm',
   contactDetails: 'Contact Details',
   error: 'Error',
