+++ packages/framework/esm-translations/translations/en.json
@@ -8,6 +8,7 @@
   "city": "City",
   "cityVillage": "City",
   "close": "Close",
+  "closeWorkspace": "Close workspace",
   "confirm": "Confirm",
   "contactDetails": "Contact Details",
   "country": "Country",
